requirements.txt
python-dotenv
google-generativeai
groq